build_image: xip: Teach the XIP component to inject additional symbols.
Since the XIP component does a linking pass on its partially linked input, it
can pass arguments to ld (indirectly by way of gcc) to inject additional
symbols.

This can be used to hard code values in that binary which aren't known at
compile time but have been figured out once the image has been placed. This
enables an optimization where we don't need to search for the anchor structure
for the dcdir, we can just know where it is ahead of time.

More crucially, when using Intel's FSP, we're required to pass it pre-baked
structures in memory which may have the addresses of other pieces of the image
built into them. Because we're initially running out of ROM and can't update
any of those values at runtime, they have to be baked in at the time the image
is constructed. Those values could be defined by edict and used as constraints
when constructing the image, but that could be error prone and leave
inefficient gaps or inadequate space for unexpectedly large components. By
using this mechanism however, we can put things were we want or where they
happen to end up, and then record that placement back into the image.

Since this behavior isn't strictly related to executing in place and is just
incidentally possible because of how it's implemented, it might be a good idea
to split the linking functionality into its own base class which XIP can
inherit from. Since there isn't anything now or in the forseable future that
needs late linking other than XIP, it seems like too much overhead to add
without enough benefit to justify it.

Change-Id: I2cada239527485f23c33bc436c49f8a74b427235