build_image: QEMU: If the dcdir isn't where we expect, look for it.
Also add a check to actually verify that the anchor is where we expect. This
will make the image more robust in case it gets modified somehow after
creation. It *shouldn't* be the case that the anchor needs to be moved, but
it would be nice if it had to be, that the QEMU entry point would be able to
handle that without recompilation.

Since scanning for the anchor is extra wasted work, we output a warning
message if that's what we have to do, so the situation can be fixed if it
wasn't expected.

Change-Id: If08dd516811907ed2ed174e07a955525ba433f03