board: Create a board_power function.
This function will return the appropriate driver structure for power
operations. This model replaces having to install a power driver ahead of
time. There are two benefits to this approach. First, it avoids having to
set up a power driver in case somebody wants to call into it later. Second
and perhaps more important, by isolating access to the power driver from
other setup code, we can avoid touching itfrom a linking perspective unless
that driver is actually going to be used. That should result in a smaller
binary if there are driverse we need for some modules but not for others.

Change-Id: I22cd0e9ae7d0cf85d390a71907391e8312d86cec