board: Add some macros to simplify defining board components.
The functions either have a board_ prefix if they're public symbols, or a
get_ prefix if they're private. There's a version for static values which
don't need to be cached, and a version for dynamic values which are generated
once and then cached for future calls.

These are in their own header file and not in board/board.h since the later is
likely to be included in a lot of places, and I don't want these macros to
leak into other contexts. Their names are purposefully short, and that makes
them more likely to collide with other things in other files.

Change-Id: I45eeacaed348f75fa5f75bbea3b1c97184821575