Delete the oak board.
This board is no longer supported by coreboot and isn't of immediate interest
to us. If we decide we want it in the future, we can always dig it back out of
version control.

Also delete drivers which were only used by oak, and a special linker option to
work around a CPU errata which was only enabled for oak.

Change-Id: Ib9e1375d7db3395585dd0ca15ddfad9ef03c1b6e