flash: Delete the unused ICH flash drivers.
It looks like these haven't been used in a long time, possibly ever? I vaguely
remember doing an experiment to see if using the SPI controller directly was
faster than using the memory mapped version, and that it wasn't any faster.
Maybe this is left over from that?

In the future, it may be necessary to write to the flash to use it to store
cached memory training data. If that happens and these drivers would have been
appropriate to use, they can be retrieved from git history.

Change-Id: I1c013ea12ae5725562f90ddf4d869c27375dc33b