uart: Change the uart drivers into depthcharge style structures of callbacks.
This merges the various 8250 drivers, makes all the UART configs default to
off, adds a board directory for qemu-x86 (so it can define board_debug_uart),
and adds a "console" class of drivers which will attach various device drivers
up as consoles.

As originally implemented, the GDB transport level code sort of pushes the
serial console driver to the side when it activates and takes over outputing
console. Moving forward, GDB will act as the serial console always, either
sending console output through directly when not debugging, or wrapping it
in the GDB protocol when debugging is turned on.

Change-Id: I7af6f786345eabde1e29e92305a3807a24acf53b