console: Get rid of some (almost) unused console related functions.
This is used by the GDB console driver to switch between itself and the uart
console driver. The ultimate goal is for the GDB console driver to be self
sufficient and to act like the uart console when not debugging, so switching
things in and out isn't useful moving forward.

Change-Id: Ic69c8e6ce5b45cf8fa18ec963a0c1898616b8ca6