Add init function priority levels, and get rid of some init boilerplate.
There were two reasons some software level work (not hardware initialization)
had been left to individual modules. One reason was that the initialization
might not always be appropriate to perform. The other was that there might be
a good reason to ensure some things happened before others, for instance to
twitch to show the system was alive first before something else had a chance
to break boot, or to ensure the consoles were registered to get log output.

To help remove some of this initialization boilerplate from all the modules,
init functions now can have priority. This lets us guarantee things will
happen in a particular order, even if they're not being called in an explicit
sequence. Also, if some initialization needs to happen based on whether or not
certain features are compiled in, then the initialization functions will
automatically either be included or excluded with that code and won't have to
be managed explicitly.

If initialization is truly module specific and not correlated to whether some
other code was included, those calls were left in place.

Change-Id: If73a5094134496f43adddafd3a6924f3fa2c1dda