head: Stop clearing the BSS manually at startup.
This should be done by the loader, not the image being loaded. I've heard that
GDB doesn't do that. Rather than work around what I'd consider a
bug/unfortunate decision in GDB in code that will normally not run under those
conditions, lets find a workaround in GDB if/when that problem arises.

Change-Id: I6c447c7b230755c566825512a16d6058c7d4675e