Pull delay and timer_us functions into base/time.[hc].
The timer_us function is generic and sits on top of the timer_raw_value and
timer_hz functions which are implemented by the timer drivers. Instead of
putting them next to each other which might be confusing, pull it out into
base/time.c, and rename it time_us. Also, create a header file called
base/time.h for time_us, and for the delay family of functions which used to
be in libpayload.h.

Change-Id: I02e854732b0aac7614f392cbba36f4d800d7af8b