Change "main" to "module_main".
Traditionally, "main" has had a weakly defined signature. It sometimes returns
an int, sometimes void, sometimes it takes argc and argv parameters, and
sometimes it doesn't. In depthcharge, it never returns anything, and it also
doesn't take any argmuents.

This change creates a well defined "module_main" signature which takes no
arguments and returns no value. If it returns, the calling code is responsible
for notifying the user and halting the system. The printf at the end of each of
the existing main() functions that handled that job have been removed.

Change-Id: I1a18f6d0df40e4a537b8cadb99b4c3182fef1cab