Disable make's built in rules when building depthcharge.
We don't rely on them, and they slow down make and make it do surprising and
incorrect things sometimes. For instance, a built in rule was making it try to
build a module which didn't exist, and that was causing spurious warnings and
delaying the build. With this change, everything still builds, the warnings
are gone, and the build at least feels a little peppier.

Change-Id: Ic5743e76feb7b22a1a74a89edbee063867739b09