Get rid of some small, unused utility functions.
The ones which would convert between BCD and "decimal" would be useful if we
were reading or writing from an RTC type device which tends to store its
values in BCD. Since we no longer have code in depthcharge to do that, BCD is
an oddity that we no longer need to support.

The bin2hex and hex2bin functions look like they'd be most useful when printing
or reading in hex values from a string, for instance to/from a console or from
a string representation of data stored somewhere. Since printf can already do
the printing part and we aren't using the reading part, lets get rid of those
too.

Change-Id: Ia312d28d6a4da43e07b40717f32b7f93f5217702