Pull small macros/inline functions out of libpayload.h into base/algorithm.h.
My first inclination was to put these in a header file called math.h since
they're all numerically based, but that name is already taken. Some poking
around online suggests that a name like this should be fairly unsurprising.
It's taken in C++ for some similar standard functionality, but since this uses
a C style file name it should be ok.

Change-Id: I7ff88b5005c58364ebdc7b1d0c844ccd9adfa02a