Rework depthcharge's earliest code and handoff from earlier firmware.
This change pulls the very first code that runs out of the libpayload area
and puts it into depthcharge. It also splits it into 4 parts.

1. Handoff specific assembly level initialization. This is an opportunity for
handoff specific state to be saved before it's lost in later initialization,
or perhaps to transition the CPU into a state later code expects.
2. Generic assembly level initialization. Currently this just installs the
new stack.
3. Generic C initialization. This does a lot of the work and handles common
startup tasks.
4. Handoff specific initialization. This is called into during the generic
initialization, and happens after some minimal hardware setup, for instance
installing exception handlers, but before most other initialization has
happened. This is so the later initialization can receive information from the
earlier firmware in case that affects what it needs to do or how it needs to
do it. That means this earlier code will likely not be able to report errors
and will need to defer them until other initialization has happened.

Legacy mode will no longer pass a parameter to later, unknown firmware
modules. We can't reliably depend on having coreboot tables to pass to them,
for instance, and the firmware database structure is not for external
consumption.

Also, since we haven't weaned ourselves off of the sysinfo structure and
coreboot based initialization yet, on ARM that structure is passed through to
later firmware modules instead, and they will re-translate it into their own
firmware database. Once we no longer rely on the sysinfo structure, that
mechanism can change over and we can use the fwdb throughout, only looking at
other sources of information at the transition point in the handoff code. On
x86 the firmware finds the coreboot tables on its own, so we can pass the fwdb
through to the RW firmware. This will break older ARM chromebooks which have
RO and RW depthcharges, but that will be fixed in time.

This change also arbitrarily chooses a region of memory for the FWDB if one
doesn't already exist. This is an instance where depthcharge could really use
some sort of mechanism to keep track of and allocate regions of physical
memory, both to keep things from stepping on each other, and to make the
process of zeroing unused memory less manual and error prone.

Change-Id: Ied8dae01e315f736fc22121fce12255f3e2215dd