power: Replace the sysinfo power driver with a GPIO reset driver.
The original power driver which became the sysinfo power driver used an
underlying driver, for instance for a PMIC, to power off the system, but used
a GPIO to cause a reset. That hardware setup became common, and the wrapper
which bound a power driver and GPIO ended up being copied into multiple boards.
Those were ultimately consolidated, but at the same time the explicit GPIO was
replaced with a GPIO generating function which used the coreboot tables to
figure out what GPIO to use.

Since we can't count on having coreboot tables, this change moves us back to
using an explicitly configured GPIO. The GPIO configuration to use was
determined by looking back in depthcharge's revision history, and looking at
how coreboot is set up to do its own resets.

Change-Id: Ibe7c684e24f96ea28a3dd006afba293c14c82c35