GPIOs: Pass GPIOs to sysinfo_install_flags instead of a generator function.
Since we can't necessarily count on coreboot having run before us, we also
can't rely on it telling us where certain GPIOs are located. Instead of
passing a function which creates appropriately configured GPIO drivers using
the information in the coreboot tables, we now pass in the GPIOs already
configured using information local to depthcharge. If a GPIO shouldn't be
resampled at run time, a NULL pointer should be passed in. Values read and
passed through from coreboot are still read from the sysinfo structure for
now, but since that information is also in the FWDB that should change in the
near future.

Change-Id: Ib21b9948e383e40a5680ca02418ac7d13a638cc3