flags: Return flags with board functions instead of installing them in a list.
If flags are collected in a list, then even if they aren't used, the code to
make them available still has to be compiled into the image. Also, there needs
to be code which runs at the start to actually install the flags, and there
needs to be checks to see whether the flags exist before they're used. Also,
the interface as it existed required flags to present a set of GPIO operations.
Only the "get" method of that was actually used, and it also meant that
non-GPIO flag providers (like constants) needed to be wrapped in a fake GPIO
object.

In the future, flags should generally be computed directly rather than depend
on values which are in the FWDB. That could be part of enabling any particular
board to run as something other than a coreboot payload, but is a bit out of
scope for this change. The FwdbGpio driver might also be useful if there
genuinely is a value which needs to be passed along and can't be directly
derived for some reason, perhaps the result of some decision made by earlier
firmware based on information that's no longer available.

This change also adds "fake" flags to the qemu-x86 board. These configure
vboot to run in developer mode under "normal" circumstances. In the future,
some mechanism may be implemented which allows custom values to be plumbed
through so the same image can simulate developer mode, recovery mode, etc.

Change-Id: I8f8a4983620b12b13add915eed3d66ea8969f7ec