dcdir: Add code for traversing a dcdir structure.
This code assumes the underlying medium is flash. An abstraction will likely
be added in the future which makes it possible to have a dcdir on other media,
for instance a disk.

Also, a higher level function might be added to stream line accessing a dcdir
region based on its "path", where components are separated by "/"s and are
looked up one at a time.

Change-Id: I06dc67fb390a0a4f00682b7bb230a1ca19abe821