fsp: Call the first FSP entry point, TempRamInit.
We're fudging things a tiny bit by making up a code region location and size.
Ultimately this should be where the next depthcharge module expects to run
from, somewhere down lower which overlaps with memory. That will allow us to
unpack its ELF as it expects, and continue to read from the actual ROM
contents and not just the cached versions.

Eventually, that module will need to call into the FSP which will then use a
continuation function to resume execution of the main firmware somewhere. It
will expect that to be at a particular address, and since we couldn't have
configured memory by that point, we can't have actually put anything in memory
for it to target. We're going to likely have to make a trip back through the
ROM and XIP code on our way to executing out of RAM.

Finally, I think we may run into a limitation of the depthcharge configuration
mechanism soon. Currently, certain parameters like heap size and base address
are shared across all versions of depthcharge. That worked fine when there
would be only one or two versions that would run, all after the system had
been mostly configured. Now that some modules may need to run while some very
different constraints are in place, for instance from a size limitted
cache-as-ram area, it may be necessary to build some modules with different
config options.

Up to now, what differentiation there's been has been from including different
.o files which would bring in different versions of common functions, for
instace the module_main function which drives most of the modules behavior.
This configuration mechanism would need to actually change how the .o-s were
built, not just which ones built from a shared set of options are included
in a particular binary.

Also, effort will likely need to be made in the fairly near term to get the
module-to-module mechanism working well. That will mean linking the trampoline
to run somewhere different from where it's loaded and copying it into place
as it's being used (to avoid loading the new trampoline on top of the old as
it runs), and also changing around the linker script so that the depthcharge
image starts life as a compact region of memory, again to better fit in
constrained environments.

Change-Id: I1df5d3dd18a749b839e2a5fb601e732d0f669c46