fsp: Add code to call fsp memory init.
Some of the code/data added in this change should be split out into board
specific files, and the SPD data needs to be picked using GPIO strapping and
not using the hardcoded data that matches my samus.

Also, the alignment of the SPD data needs to be forced to something small (at
most 16 bytes) because gcc will otherwise set its alignment to 64 bytes,
presumably to match the size of a cache line and to optimize accesses to the
buffer. Because this code is XIP and needs to be linked twice, once to size it
and once to actually place it, it breaks if the offset in the image forces the
second linking to move the loadable section to different addresses than it
will actually be loaded at.

In the future, the image layout tool needs to support setting alignment, and
the alignment of the loadable section in the XIP ELF file should be propogated
to the XIP image region. Until then, small fixups like this may be necessary.

Change-Id: I63c743a233a7531302ce83b81f47c750fa58dee6