fsp: temp_stack: Move most temp_stack functions into arch/x86.
These functions wrap generic functions in preram and so should be available to
anyone using x86. There were a few functions which were more specific to FSP
which were left in place.

Also, since there was only one function which was purely in assembler, it was
pulled into c to be consistent with the other wrappers.

Change-Id: I79ca80efc4b6d065653e75d422fb0477f9ad980a