fsp: Move the stack to RAM as soon as it's available.
Before we start executing out of RAM we need to put the cache back into its
normal operating mode. To do that, we need to stop using the CAR stack and
move into RAM.

The only information we need to transition from the CAR stack to the RAM stack
is the pointer to the HOB list returned by the FSP. The HOB list itself is
located in RAM, and so only the pointer needs to be preserved. Also, to avoid
having to look it up again, we pass along the location of the FSP information
header.

Change-Id: I6751021250279d2d185fddafc450865122636aca