fsp: Call TempRamExit after switching to an in RAM stack before anything else.
While the CAR stack, etc., is active, the cache isn't going to behave in a
normal way. To be able to execute from RAM and to be able to take advantage of
the cache in its normal mode of operation, we need to run TempRamExit. Once
that's done, then we can move on to running other modules of depthcharge which
aren't XIP.

Change-Id: Ia585be217e8347ff79ea377f948f07674f4b35a0