fsp: Add some infrastructure to insert SPD blobs into the image.
Also change the seperator in the ecs argument to ",". The SPD blobs need some
sort of label to indicate what they should be called in the image, and it
makes sense to join label/file pairs with ":" and not as much with ",". That
means "," can be used to separate the pairs themselves, and for consistency's
sake the ecs argument was changed to match.

Change-Id: Id66a3876ccd663d32b9fd5b1090e546bd8b29df1