spd: Add a pattern rule for building SPD blobs from "source".
The SPD blobs are, at least in coreboot, stored in files with comments and
byte values in hex with spaces (and occasional newlines) in between. This
change adds a make rule which uses bash to munge those files into binary blobs
which hold those hex values. This bash snippet was lifted from a similar make
rule in coreboot.

Change-Id: I4e1be83b1526e415fad8763a66a19f2b42b2e918