image: spd: Add a mechanism to the makefile to handle SPDs.
Unfortunately we can't use the build system's classes or special-classes since
unless we put those in the root of the hierarchy we can't be sure they'll be
available when going down arbitrary branches of the Makefile.inc tree. Instead,
we define a make variable/function which can be lazily expanded (I hope) and
which will add a label/file pair into the system.

The binary SPD files will be added as dependencies to "modules" to ensure that
they're built and available to build_image. The pairs themselves will be
composed into an argument to build_image to tell it what SPD files to put where.
If there are no SPD files, the mechanism should gracefully exclude the --spds
option for layouts that don't support spds.

Having a semi-official global flag like --spds is a little bit questionable,
but the overhead of making the option name overridable didn't seem worth it
for now.

Change-Id: Ibb2350d91d70ee62e4f97ea89e717a404c63377a