elf: Add a cast to pacify the compiler when building with 64 bit pointers.
On systems which are 64 bit (currently aarch64, potentially amd64), pointers
are 64 bit, and we can't cast directly from the 32 bit entry point value in a
32 bit ELF header into a function pointer to call to enter the expanded ELF
image.

In reality we shouldn't even be attempting to enter a 32 bit ELF on one of
those systems since presumably we'd be running in a 64 bit mode, but since
that code isn't yet specialized in that way and we don't have any support for
64 bit ELF files currently, we'll just fudge things a bit to make the compiler
happy. That code shouldn't yet be called on a 64 bit system, so we're safe for
now.

Change-Id: Id205608c432d4c3c359e0fab0d12cb5c61ee9fb7