inttypes.h: Let the architectures define these macros however they need to.
The reason these macros exist in the first place is because they don't always
equal the same thing. amd64 needs to use just longs (instead of long longs) to
represent 64 bit values, so that at least needs to change. Otherwise the
formats are the same as they were in the original header. If they see wider
use, it would be worthwhile going through them and making sure they're all
correct for the architecture they're in.

Change-Id: Icc21e20dd3237a1dd006d148ae09d8cd9e1a153e