exceptions: Rescope exception.h and its arch versions.
The content of the architecture specific exception.h files weren't simply
different versions of the same things that could be used interchangably by
code which didn't know what arch it was running on, so there's no reason to
put it in the "arch" include directory. This change moves it into the base
architecture source directory, and the code that needs it can refer to it
explicitly instead of relying on the include path to patch in the right
version.

Also, the actual generic exception.h header is pulled out of libpayload and
put into base, and a number of small style problems have been addressed.

Change-Id: Ic20425948dca55b075b18e4013f060e6a5336d04