lzma: Merge lzma into depthcharge and remove it from libpayload.
Also clean up some style problems. The lzma code came from elsewhere, and
it's a bit opaque. I think it's written in a way that's supposed to be very
fast at the expense of being clear, but hopefully it can be reworked to be
both at some point in the future.

Change-Id: I4b680b206230f41964202743b6b0f8f6101c0a1b