lz4: Merge the lz4 code into depthcharge and take it out of libpayload.
Unlike the lzma code, this code is moved basically as is. It's only used in
one place so if it was broken it would be harder to notice, and its
interaction with its "wrapper" is more complicated, including warnings about
inlining and performance.

In the future, especially if this compression scheme sees wider use, would be
to characterize its performance and see if it can be cleaned up without
slowing it down.

Change-Id: I9ea5fddf6485868fd1384d8961b15841a82fbc1c