UEFI: Move the UEFI headers to the "uefi" directory at the source root.
Also modify all those includes to use a "uefi" prefix when referencing other
includes. That should allow making the UEFI headers universally accessible
without having to worry about name collisions. The headers generally start
with capital letters and so should be in their own namespace by coincidence,
but that's not safe to depend on.

Now that the headers can be referenced universally, they don't need special
support in the UEFI module Makefile.

Change-Id: I557f01f71f2ad240b99cf11ab0ef436b07091b2b