IO: Remove obsolete (read|write)[bwl] accessors, and fix io.h for amd64.
Those obsolete accessors were left in because they were pervasive in the code
base, and because the order of the write versions arguments had switched in
write(8|16|32) version. This change replaces all the calls to the old
accessors with calls to the new ones, removes the old ones, and fixes up a few
places where the stricter type checking of the new accessors caused the
compiler to complain.

Also, this change modifies some of the types used in x86's io.h so that
they're fixed in size and compile correctly on amd64.

Change-Id: Id8d28a6bbc2486d3d282564f795bcf0f776a54f0