TPM: If CONFIG_MOCK_TPM is turned on, stub out the TPM functions.
CONFIG_MOCK_TPM tells vboot to pretend the TPM is there and gives back all the
right answers to boot. If that option is used then nobody is going to try to
call the TPM callbacks, but with the options being used for UEFI, if there are
unresolved symbols then the build will fail. We can compile in stub TPM
functions instead which will never get called, but will also not make their
own calls to TPM related driver functions.

Change-Id: I1e72ecb82d9160e4d4b0b38624a6418135ce8c71