amd64: Add the stack size as an immediate, not as a memory address.
There was a bug in the assembly portion of the amd64 handoff code which
switched to depthcharge's stack from whatever had been in use previously.
When the stack size was added to the start of the stack buffer to find the
top/starting position for the stack, the "$" was missing, so it was using that
location as a pointer and adding in the value stored in memory at that spot.

Change-Id: Ied3a2c12619aaa1e4f5dc4264a93b8bf3eda473c