UEFI: Build the ELF version of UEFI code as a PIE and not an SO.
There are (at least) two different ways to build a relocatable image using
gcc. The first, which was used by gnu-efi and inherited into depthcharge, was
to build using the -fpic option, and to link everything together into a shared
object using -shared.

There are three main problems with that approach. First, unresolved symbols
are basically ignored in that sort of build because undefined references could
be resolved when the library was loaded. Since we aren't using it that way,
they just would point out into space and cause the program to crash. This can
be worked around somewhat by passing gcc options that make it much more
uptight about unresolved symbols, but that's not a great solution as described
below.

Second, because all globally visible symbols in the ELF are considered
possible entry points from not yet available code, section garbage collection
is hobbled significantly. Special care can be taken to hide the symbols that
aren't actually supposed to be entry points, but I couldn't get it to work,
and that's not a generally practical/scalable solution.

Also, that would leave code in the binary which would never actually be
called, and which had unresolved symbols inside of it. For instance, if the
vboot TPM callbacks will never be called because vboot has been configured to
fake out TPM interactions, it would be reasonable not to put TPM access
functions in the binary. It might not even be sensical to try to add some
since no amount of magic will conjure a TPM from nothing. But because the
callbacks were left in and "might" be called, then theree are unresolved
symbols, and the stricter unresolved symbol option above would fail the build.

Finally, building as a shared object injects a bunch of shared object
supporting mechanisms like a PLT which are not actually useful for an EFI
binary.

Fortunately, the second way to build a relocatable image is to built it using
the -fpie flag and to link it with -pie. That creates relocations which can
be grouped into a section surrounded by helpful markers, doesn't add in some
of the dynamic linking clap trap, and will clean up unused code successfully
like the non-relocatable version of the build already does.

This also trims down the relocation code a bit, and the UEFI linker script.

Change-Id: Iff8246c4d6c899408c9fe16ee4fdfed35f302a40