uefi: Use uefi.h and uefi.c.
Use the header for getting generic uefi "stuff", including system headers.
This level of abstraction avoids having to use the edk part of the path, while
still keeping the edk stuff seperate and in it's own container. It also avoids
having to know what the edk header file is called.

Use the new uefi_system_table_ptr function to retrieve the system table
pointer, either from a static variable cache, or from the fwdb if this is the
first time it's being accessed.

Change-Id: Iec353eb1b470ec2feb7319cf3d5122aac58ec0dd