storage: Add a storage driver type.
This type of driver abstractly represents a piece of storage and allows
writing into it or reading from it. They can be layered on top of each other
to represent raw media combined with potentially hierarchical layout drivers
which will make it possible to generically refer to a particular piece of
information/region of storage without having to know what media the
information is stored on, or what scheme it's partitioned with.

Change-Id: I848274c8a749280122d745ba6ad59eb3ead0a2c5