layout: Move all the other boards to use the coreboot layout driver.
This is even when it's not appropriate because those boards will use some
other mechanism when one is available. If a board could be hooked up
reasonably then it was, and if it couldn't (uefi and qemu, particularly)
then I just set them up as an 8MB flash chip and the coreboot layout.

This will get all the boards building when board_storage_gbb is used, and
won't be any more broken than it was previously.

Change-Id: I898e57f573a1d53a756332aff595c4b56f55b37c