vboot: Clean up the flash based nvstorage code.
As far as the interface, a function had a typo in it (offet vs. offset), and
the other was unnecessary. This change also cleans up the code a lot. This is
in anticipation of switching it from explicitly accessing flash to using the
storage driver interface. Before that happens, there needs to be some support
for erasing, perhaps in the write function of the flash storage driver.

It might be possible to also integrate the wear leveling features in this code
into the storage driver itself, possibly allowing folding all the nvstorage
callback implementations into one.

Change-Id: If37efd4d0fd8b05c411947b0215094a479ed34c5