storage: Implement the flash storage write callback.
This will check to see if it's necessary to erase, and also do a read/modify
write on any sectors that do need to be erased and which are only being
partially overwritten.

I don't have hardware which would actually write to flash since that's not
used in depthcharge except on a particular machine. It will be used at some
point on samus, but I'll have to reintroduce the PCH SPI driver first which
I'm not ready to do yet.

Change-Id: Ica8dfefc1fbe9d87291ed3e8c2f20a41bda500b5