dcdir: Add the dcdir_open_dir_raw function.
This function also collects information offset/size information about a
directory. The current dcdir based image setup uses a dcdir directory to hold
all the bits that are verified by vboot. That's very useful, but it means that
when verification is done, it needs to be done on the whole directory, not on
an individual region in the directory. To accomodate that, we need to know the
extent of a directory and not just have access to the leaf regions within it.

Change-Id: I080d4220c215bad3ce184463a5f8526c3e1cbe0e