crossystem: Calculate the FMAP location instead of relying on the global FMAP.
The current code makes several assumptions which are not generally true.
First, it assumes there's a global FMAP somewhere to pass along to the OS.
Second, it assumes that the FMAP is mapped into memory, and that its address
can be retrieved. Third, it assumes that the flash_read function will return a
pointer to that actual location, and not a copy of it which will be wiped out
when the OS is loaded.

Since the information is being passed in a structure which also assumes there
must be an FMAP, instead of trying to correct these assumptions, lets just
embrace them there and calculate the location the FMAP would be in on an x86
system with a memory mapped flash.

Change-Id: Id668a6e717509ee9a899ad06bdc9fd08b8ede164