memmapped flash: Change memory mapped flash into generic memory storage.
Memory mapped flash basically wraps a region of memory as a flash device,
with limitted capability, and then the flash storage driver wraps it again
as a storage driver. Instead, we can turn it into a generic memory backed
storage driver which will represent memory mapped flash in a simpler way, and
also make it available for other memory backed bits of storage. For instance,
when flash isn't available, the GBB could be loaded into the RO firwmare
binary as part of the image and then wrapped in the memory storage driver.

To make it easy to set up a memory mapped flash device which is very common
on x86 devices, a utility function was set up which wraps the region of memory
at the end of the 32 bit address space using the CONFIG_IMAGE_SIZE_KB config
option for placement and sizing. This change also goes back and sets that
config option for boards that weren't yet using it.

Change-Id: I760d259b2a3804dc6395902f74e0223c5a69b27d