trampoline: Pull the binary wrapping logic into a helper script.
That script can be used to wrap other binary data for use in executables. For
instance, it could be used to wrap the "read only" image contents and embed
them in the "read only" firmware as used on a UEFI version of depthcharge. I
put "read only" in quotes since it won't actually be read only, it will just
be potentially verified by UEFI using secure boot before being executed.

Change-Id: I6c5780e1edd4a2e72e7da80b7eca1d5f4fbd2366