imagelib: In the Area classes build function, use the computed_min_size.
It's possible that the outermost Area being built is not fixed size. It
doesn't make sense for it to be expanding since it could expand infinitely,
but it could shrink to minimally hold its contents.

To support that case, we should pass the computed_min_size down the hierarchy
when placing the main Area's children, and not _size which is only set if that
area is of fixed size. If an Area *is* of fixed size, then the two values will
be the same and there should be no change in behavior

Change-Id: Id98c1d514a088bc8b906cfe3e11a31f625f56887