layout: Add a UEFI layout driver.
This layout is mostly the same as the dcdir layout driver, except that it
expects three different underlying storage drivers for the RO and two RW
firmware areas. There are also no common RW areas defined, and no place for
EC images or their hashes.

This is consistent with how UEFI "images" will be stored. The RO part of the
image will be embedded in the depthcharge application itself, so that whatever
mechanism validates the "read only" depthcharge will also apply to the RO
image contents. The RW parts will come from other files which are stored
somewhere else, and are not validated by UEFI since they'll be verified by
vboot and need to be updatable without having to add a UEFI signature to them,
or routing them through a mechanism which would validate that signature.

Change-Id: Ic157f6b95b9f33a3c7edf73fe51e73caec638a35