UEFI: Add a protocol header for the EFI shell.
This is distinct from the EFI shell parameters protocol, and provides a lot of
callbacks for accessing files. The raw file access mechanisms in the base UEFI
specification are complex and hard to use, and since it's probably mostly safe
to assume the shell is supported on the devices we care about, it's better at
least for now to assume that it's there than to try to build something out of
the primitive mechanisms.

Change-Id: I3c47d4851979a9c943a8598fcc6e9585f3f63794