UEFI: Accept two command line argument paths and add them to the FWDB.
These files are the RWA and RWB images associated with the current RO version
of the firmware. Originally I'd hoped to keep them around as files that could
be read/written on request, but I'm not sure that would actually be used or
that it would actually work in the general case where the RW images may have,
for instance, been loaded off of network storage somehow.

Change-Id: I3e26f9d971dbbe9e6a44a25a45e8fac4df985c86