storage: dcdir: Fix how the media pointer is passed down the hierarchy.
The "media" pointer was being used both as a way to indicate that a structure
was initialized, and as a way to refer to the underlying media when a storage
area was read or written. It was also being passed down the dcdir storage
hierarchy until it got to the root, at which point it would be initialized and
make its way back up the tree.

Unfortunately, if an intermediate node was already initialized, ie it already
had its media pointer set, it wouldn't call any farther down the tree and
would never get to the point where the media node was initialized. Since each
node in the tree now has a media pointer anyway, lets just make setting it
from the parent's copy part of initializing a child. That way it will happen
unambiguously, and can still act as a marker for how far down the tree you
need to go.

Change-Id: Id2f0ad32d881f4e7265c64938aecdf0df89d695f